+++ lib/features/repository/presentation/ui/repository_screen.dart
@@ -244,6 +244,9 @@ class _RepositoryScreenState extends State<RepositoryScreen> {
                   _repositoryBloc.add(UpdateRepositoryStatus(status: RepositoryBlocStatus.initial));
                   break;
 
+                case RepositoryBlocStatus.fetched:
+                  context.read<BranchesBloc>().add(GetRepositoryBranches());
+                  break;
                 case RepositoryBlocStatus.repositorySelected:
                   context.read<BranchesBloc>().add(GetRepositoryBranches());
                   _workingDirectoryBloc.add(GetRepositoryStatus());